Provide ghe_create_data_dir parameter to ghebackups::config manifest.
The ghe_create_data_dir parameter is the only one not provided to the
ghebackups::config class declaration in the init.pp.

With future parser off, all values, including ghe_create_data_dir are
set properly in the template config file.

With future parser enabled, the default value 'yes' for
ghe_create_data_dir is not parsed. This results in an empty string in
the config file that ends up on the node.

If you try to specify a value in the ghebackups declaration, you end up
getting the following error.

Error: Invalid parameter ghe_create_data_dir on Class[Ghebackups]

Looks like the right thing to do is specify the ghe_create_data_dir
parameter the same way the other template parameters are done.